1. Champ Base 오브젝트 생성 2. Archer Champ 오브젝트 생성

+++ 660Engine/sszStatIcon.cpp
@@ -43,6 +43,6 @@ namespace ssz
 	}
 	void StatIcon::SetIcon(eStatType eType)
 	{
-		GetComponent<MeshRenderer>()->GetMaterial()->SetTexture(mStatIconTex[eType]);
+		GetComponent<MeshRenderer>()->GetMaterial()->SetTexture(mStatIconTex[(UINT)eType]);
 	}
 }
